move test helpers for pytest command line
