+++ setup.cfg
@@ -1,0 +1,2 @@
+[tool:pytest]
+norecursedirs=tests/helpers
